dts: pmi8940: disable PWM backlight dimming
The use of PWM (pulse width modulation) to dim the display backlight at
lower brightnesses is not desired because it can cause eye strain and
headaches. It can also make the display look less smooth. Our backlight
LED seems to be able to handle lower analog dimming (by voltage and/or
current), so use it solely.

Signed-off-by: kdrag0n <dragon@khronodragon.com>